Try harder still to get a local host struct hostent
Will Cohen came across a failure to start pmlogger running locally
[Mon Jan  6 15:13:24] pmlogger(23927) Error: __pmGetAddrInfo(santana), Unknown host
[Mon Jan  6 15:13:24] pmlogger(23927) Error: Can't get host name/IP address, giving up

It looks like this is another case of failure to resolve local
addresses from local host names.  Documentation for getaddrinfo(3)
states a null can be passed in for the "node" name (first param),
so we now use this in cases where we know we're dealing with local
name resolution, as a fallback.  Audit found a second similar-ish
case over in pmcd, dealing with socket PMDA connections - make it
use this same mechanism instead of hard coding "localhost".