Updates to pmfind after further review and discussion
Reading through Daves recent man page additions, a few things
cropped up related to the pmfind tool.  The -d ("domain") option
was not yet documented, nor used in the tool, so added those in.
Then realised the term "domain" was fairly ambiguous (networking
domains and PMDA domains); so changed it to "mechanism", as used
in the pmDiscoverServices man page already.  And -d become -m as
a result.

Added in the -q option to pmfind (quiet mode), and added error
handling to the pmDiscoverServices call.  Allow alternate service
names to be given, preparing for future.  Added -D (debug flags)
support into pmfind too, while I was in the neighbourhood.