pmdagfs2: Add complete set of GFS2 tracepoint metrics
Added the complete list of GFS2 tracepoints as found at: /sys/kernel/
debug/tracing/events/gfs2/ 82 new metrics have been added along with
14 new control metrics (96 total new metrics).

The metrics for the following GFS2 tracepoints have been added:
    - gfs2_glock_state_change
    - gfs2_glock_put
    - gfs2_demote_rq
    - gfs2_promote
    - gfs2_glock_queue
    - gfs2_glock_lock_time
    - gfs2_pin
    - gfs2_log_flush
    - gfs2_log_blocks
    - gfs2_ail_flush
    - gfs2_block_alloc
    - gfs2_bmap
    - gfs2_rs

The glock_lock_time metric has been renamed to worst_glock as this
more closely represents the stats that the metric provides (and so
there are not two metrics with similar names). The new tracepoint
metrics are handled in the ftrace.* files and the newly renamed
worst_glock.* separately handles the worst_glock metrics.

With these changes there has been slight tweaks to the update the
readability of some comments and the removal of some unneeded
includes. Help, QA (654, 655 and common.gfs2)and the Readme have all
been updated to reflect the changes in number of metrics and the name
change for the worst_glock metric.

QA tests and Covscan has also been run upon the code changes.