pmlogger PR1041: implement limited duplicate-elimination amongst metrics
Duplicates are detected during parsing, in the dometric() function
which fires after a new metric & instance-spec is identified.  It is
skipped if the same metric & group-specs & non-instance-specificity is
detected in a prior task_t group.  This dramatically reduces the amount
of logging data generated from a default-pmlogconf running against some
random Linux box (46.36 -> 17.01 Mbytes/day).