Optimisations to the pmlogger task_t fetch groups
This change builds on recent improvements and optimisations to the
pmlogger fetch mechanism.  Earlier we tackled squishing of metrics
specified multiple times, down to one, reducing the generated log
size in this case (significantly!).  We now go further, working on
producing more optimal internal data structures in pmlogger as the
configuration is parsed.  This has the effect of also removing the
duplicate metrics, but now adds duplicate instances, and even more
importantly perhaps, reduces (signficantly) the number of pmFetch
calls required to satisfy the requested configuration.  This gives
us some on-disk size improvements, as well as reducing the work on
the server (pmda/pmdas).

Internally, the lifetime of the task_t data structure changes.  We
used to allocate a new task for each configuration block, even if
they specify the same interval/state, registering each immediately
with the asynchronous fetch mechanism (each resulting in separate
pmFetch calls).  Now, we allocate only one task_t for any config
block that has the same interval/state specification, and we delay
the registering of the asynchronous fetch callback until parsing
is complete.  In other words, we delay decisions about what PMIDs
and instances will be in each fetch until the last moment, which
allows for optimal data structure generation at that point.

The global tasklist is a singly-linked list of task_t structures,
which is constructed by prepending new task_t's to the list head.
One non-obvious result of these changes is that due to delaying
the registering of asynchronous fetches, they are now registered
in reverse order to before.  This has caused one QA casualty as a
result of different pmResult ordering causing different numbers of
log reads (qa/465).  I've verified that changing the order of the
two blocks in qa/src/config.interp gives the previous behaviour,
so this test was always accidentally relying on config file order.
This is the only remaining QA failure - it will be resolved via a
separate commit.

Test qa/061 is updated to filter the libpcp warning message about
the AF event queue being empty.  This is because we're collapsing
tasks more effectively now - in those cases, the event queue used
to have a no-op task added, which is no longer added to the queue
at all, hence we no longer trigger the message (so filter it from
expected .out so older pcp variants also work with the test).

Test qa/248 is updated to remove the recently added debug-diags to
reflect when metrics were being dropped.  Code redesign meant such
a diagnostic no longer makes sense.

Test qa/418 is updated to sort the metrics reported via pmdumplog,
as the may (will) now be logged in differing order, so test output
became non-deterministic.

New version of qa/465 output demonstrating fetch optimisation (one
pmResult logged only now!).  Similarly, new 466 output.

Test qa/633 is updated to give pmlogger -s1 instead of -s2 - again
the optimisations result in one fetch not two satisfying the given
configuration file.

Adds new test qa/726 with several of the trickier cases of instance
and metric de-duplication that happen with the new pmlogger parsing
logic.