Workaround pmlogger qa archive generation issue in interp logs
An earlier commit (a7f30720) mentioned...
"in reverse order to before.  This has caused one QA casualty as a
result of different pmResult ordering causing different numbers of
log reads (qa/465).  I've verified that changing the order of the
two blocks in qa/src/config.interp gives the previous behaviour, "

This incorrectly referenced test qa/465 - test qa/087 was intended.

This commit words around the problem by flipping the configuration
order of the logger config.interp metrics, ultimately so that they
are logged in the same order as always.  Open to better suggestions
for fixing this - an earlier idea using logextract turned out to add
more problems than it fixed.