Fix removal of plots from a chart with multiple selections active
The way the selected items (metrics/plots) were being removed from
the hierarchical list of metrics-in-a-chart on the Edit Chart window
did not always work correctly - only *some* of the plots were culled
when multiple plots are deleted at the same time.  This was a result
of the way the iteration was being performed.  Resolved it by going
a different route, first getting the list of all selections and then
removing each one rather than removing (hence unselecting, affecting
the iterator) as we walk the list of selections.