Resolve pmdammv sigsegv when no MMV tempdir present
There were situations where we could call into the PMDA
tree code (dynamic namespaces) with a NULL pmns pointer,
which it would dereference.  Add checks to that code (as
other tree API calls have already) and ensure pmdammv no
longer calls it in this anti-social way.

QA tests 213 and 646 updated to function with a missing
MMV temporary directory as well.