Insert nfsclient PMDA into the build (not yet packaged)
As discussed on-list, Martinns is interested in continuing on
with Max's initial review on pmda nfsclient.  Pending review,
tweaks and a QA test or two, also as discussed and originally
suggested by Max, this is built but not packaged yet.

Have fun, and a belated thank-you for the contribution Ben!