Allow PMDA Install scripts to be run even when pmcd is stopped.
Following the Install recipe from the GFS2 PMDA man page, users
can run into difficulty if pmcd is not running first.  There's
no mention of that, and the error message is cryptic.  This fix
extends the generic PMDA installation process to allow any PMDA
to kick-start the pmcd process if it needs to.  The one wrinkle
there is a chicken-and-egg with the namespace setup, but that's
tackled via addition of a Rebuild call in the right spot.

Reported by Andrew Price, tested by Paul Evans, myself and now
also PCP QA test 755.

Resolves Red Hat bug #1062443.