Changes to pcp.conf packaging to support "multilib" packaging
Move the configuration files (pcp.conf in particular, as it is
not architecture neutral) from pcp-libs into a new pcp-conf
package packaging formats for both rpm and deb formats.

As this is a fairly major packaging shift that people will at
times probably need to refer back to, bump the minor version
number (3.8.x -> 3.9.0).  A dependency exists between pcp-libs
and pcp-conf of course, but this is no longer arch dependent.

This is part two of three for rpm "multilib" support for pcp,
where both 32 and 64 bit variants of the -libs and -libs-devel
packages can be simultaneously installed.

It also happens to resolve a long-standing Debian bug, so its
got that goin' for it, which is nice.