pmlogextract - <mark> record handling fix
Previously, if pmlogextract was selecting a subset of the metrics
from the input archive, then <mark> records from the input archive
were not copied to the output archive.  This could have lead to
incorrect value interpolation in the temporal region where the
<mark> record should have been.

Most uses of pmlogextract do _not_ use metric selection, so this
bug is not as serious as it may seem at first ... debugging interp.c
and QA are the two places where we were exposed!