Fix valgrind errors in pmdalinux /proc/stat refresh
This one has bugged me for a long time, and I finally needed to spend
some time hunting an obscure memory corruption in pmdalinux - so got
to revisit this properly.

There were a few small problems in refresh_proc_stat causing valgrind
to report "invalid read of zero size".  Most problematic was the code
reading the stat buffer initially, which slightly confuses the offset
and size combo in passing a buffer pointer to read(2).  The realloc
call here also happened to allocate 8x the (small) memory it wanted,
accidently multiplying the size by sizeof(char).

The nbufindex calculation also ended up with one more line than was
in the buffer, which was OK because it always created an empty line
at the end of it (which doesn't actually exist in /proc/stat).

Finally, added some small improvements to the handling of ENOMEM for
these global stat buffers, and removed a double-memset on the initial
call (which callocs a zeroed buffer).