Fix s390x platform issues in /proc/cpuinfo parser
The format of /proc/cpuinfo on s390x is nothing like any other
platform.  We cannot attempt to push it through the same code
path, it simply has nothing in common.  If we try (as we used to)
things fall apart when we use the (negative) cpu number local
variable to index into the cpuinfo array.

Resolved by adding in some defensive code to short-circuit the
decoding logic when an unexpected file format is seen.