Do not include the generated qa_outfiles in the source tarball
This excluded qa_outfiles (generated list of QA test validated outputs)
from the source tarball shipped for each release.  The problem that its
caused is that patching a tree with a new test results in the test .out
file being exluded from pcp-testsuite, since its not included with the
original list.

Resolves Red Hat bug #1064311.