pmie - quieten warning after tmp dir perms change
After recent changes to the mode and ownership of the
$PCP_VAR_DIR/tmp directory, the message "... warning cannot create
stats file dir ..."  may be emitted each time pmie is run as a
user other than "pcp".

This happens a LOT in QA.

Since the message is only a warning, and the only side-effect is
that the pmie process is not visible in the instance domain of
the pmcd.pmie metrics (it is likely that no one but kenj cares!),
I've suppressed the message unless one of the pmie verbose flags
is set (-v, -V or -W).