pmie - another day one bug, this time in count_* operators
The operand of the count_inst, count_host and count_sample operators
is a logical expression.

If the expression is set values (multiple instances for count_inst,
multiple hosts for count_host, etc), then the code did not check for
the tri-state value of UNKNOWN (or DUNNO internally), rather it added
the values assuming 0 for FALSE and 1 for TRUE ... DUNNO is 2 which
explains why the result was TWICE the size of the instance domain
if the expressions was undefined over and instance domain.

Part 2 of the bug Chandana de Silva discovered with this simple
pmie rule:
    count_inst( match_inst "httpd" proc.psinfo.pid > 0)  > 0
	-> print "count %v";
that reported 6 most of the time, and sometimes reported 1400+

qa/815 now exercises something similar.