Handle unix domain socket binding (creation) for pmlogger when run as a normal user.
When running as a normal user, pmlogger is unable to bind to
(i.e. create) a unix domain socket in /var/run/pcp (PCP_RUN_DIR),
because that directory is not writable by normal users. This change handles
the situation in a similar way to how the port map file in
/var/lib/pcp/tmp/pmlogger (PCP_TMP_DIR)/pmlogger is handled, which is to
continue on an to issue a message only if DESPARATE is defined
at compile time.

This solution is reasonable, since pmlogger will go on to create
inet and ipv6 sockets for control connections. Only if all
sockets fail to bind is this a critical error.

This addresses the regression of qa tests 039 046 053 085 161 178
234 242 248 304 352 354 466 524 593 and 650.