More testing of cases from Miloš and a final take on pcpqa user setup
Further combinations of existing system state (in terms of which dirs
already exist and which do not) uncovered further pcpqa account setup
issues.  We now take the -M approach of not having useradd create our
homedir, just as we do for the pcp account, knowing installation will
be creating it for us (take that responsibility away from useradd).

Resolves Red Hat bug #1025688.