Add pmchart options to set font size and family
Introduce -f and -F to pmchart for allowing the user more
control over the default chart font size used for things
like Y-Axis labels, the title and the legend.

Other excellent suggestions for improving this include an
interactive Preferences dialog addition, and browser-like
Ctrl+ and Ctrl- hotkeys.  These have not been implemented
at this stage.  Additional work on improving layouts with
these settings would be good too, but this gives at least
a simple way to tweak this now.