pmlogger: add -z option for host-local timezone support
It is difficult for a pmlogger user to specify an absolute time at
which she'd like the process to shut down.  That's because the
pmlogger -T option is interpreted with the remote PMCD's time zone in
mind, which could be an a-priori unknown.  This patch adds a -z
option, which lets pmlogger interpret -T / etc. in the originating
local time zone instead.

pmmgr's pmlogmerge-granular mode is updated to use this.