Improve pmchart metric selection with multiple hosts
For the multi-host case, do not expand all namespace trees
all at once in the New/Edit Chart dialogs, as it results in
user interface clutter.  Just show the top level hosts, at
first, so the user can choose where navigation should begin
and so they have a clear view of available hosts.

When many hosts (5+) are in use, this becomes a massive aid
to the metric selection process.