pmmgr signal handling: setpgrp
Shutting down an entire pmmgr family of processes is tricky, because
we use process-group-based signalling to forward SIGTERM to everything
we started.  But in some cases, pmmgr may not be in a process group of
its own, such as if it is started by init scripts, instead of by hand
from a shell running on a pty.  To make both situations similar, we
now run a setpgrp() near initialization, to be sure that pmmgr's own
pid can be used (negatively) as a process-group.

pmlogger has had a hardcoded conflicting setpgrp inside it, for its
normal use as a daemon.  However, when invoked by hand, or via pmmgr,
it now considers itself in the "foreground", and does not setpgrp away
from the inherited one.