Add preference setting save/restore for saved hosts and fonts
The concept of a "Saved Hosts" list is introduced into pmchart.
This list is intended to be useful for remembering an oft-used
set of hosts (perhaps a local cluster) such that the hostnames
to not need to be added via command line or interaction every
time pmchart is invoked.

This list can be interactively modified (hosts added & culled)
using the Hosts tab on the preference dialog.  A new -H option
(using pmstat convention) is added to allow the list to be set
using the command line too.

Some work for remembering the users prefered font setting is
also included in this commit, bit too intertwined to extract.

Resolves Red Hat bugzilla #1069947.