libpcp/interp.c - refactor code
Relax!

This is purely cosmetic to be more transparent about how the
state of observed values and <mark> records is managed during
interpolation.

There is no functional change here and with these changes, all
related QA still pases.

This is a new code baseline that will be used to fix the bad
performance that Frank has observed in archives with lots of
<mark> records.