Add support for a global font preference setting
The Font preference tab is now fully functional - this
tab allows one to modify font family, style and size on
the fly and to remember that setting for future pmchart
invocations.  It has immediate effect.

Some additional work is done in preference save/restore
code to clear settings if the match the defaults.  We
should do this kind of thing for the other preferences
also, but that is left as an exercise for another day.

Resolves Red Hat bugzilla #1069943.