configure - static probes for freebsd
The existence of <sys/sdt.h> is not enough fo FreeBSD ... the header
can be there, dtrace can be installed, but if the kernel is not
configured in the right way, dtrace -h fails and with it goes the
build.  Be tougher before deciding that HAVE_STATIC_PROBES should
be set.