Fix a python module bug for PMDAs without instance domains
The interface to the pmdaInit(3) routine in the python wrapper
code was not correctly initializing all fields.  This meant in
the case of a PMDA with no indoms (nindoms == 0) we passed an
uninitialised buffer pointer in.  This should be harmless, but
the pmdaInit(3) code is very strict in checking its parameters
and refuses to proceed in this case when the buffer pointer is
not NULL.

The actual text of the error produced is:
  "pmdaInit: PMDA xxx: indoms not consistent with nindoms"