Remove pmchart one-metric-multiple-hosts legend ambiguity
This change tackles the visual ambiguity that exists when a
chart is created (or edited) such that the same metric from
multple hosts is used.

A number of difficulties arose in attempting to retrofit the
prepend-hostname-to-legend-labels policy (potentially having
to undo later on when editing occurs too) after a subsequent
plot is observed to introduce the ambiguity.  So, in the end
went with the far simpler solution of always using shortname
-prefixed label for plots (all this is only when no explicit
plot labels are specified, of course).

Resolves Red Hat bugzilla #1069960.