Fix __pmConnectLogger() access control bug for the "local:[path]".
When pmlc uses local:[path] is specified, __pmConnectLogger()
should attempt connection as "unix:[path]" and, if this fails,
it should try again as "localhost".

This was working ok for the case when connection via unix:
could not be established. However when the connection was established
but terminated due to insufficient access control permissions,
__pmConnectLogger() did not go on to try connecting as "localhost".

Also, -ECONNREFUSED was being returned in this case instead of -EPERM.