pmdagfs2: Miscellaneous updates
Corrections and slight updates to the text contained in the
help text for the PMDA. Spelling and Grammar issues corrected
and expansion of the gfs2.tracepoints.rs.* help definitions.

Removal of a small code formatting niggle in pmdagfs2.h

Removal of log printing testing code from worst_glock.c that
made it into the last release.