Post-review, pre-merge cleanups for glib getdate use
Packaging builds fail (as mentioned in earlier review ;) with this:
gmake[5]: *** No rule to make target `getdate.y', needed by `getdate.c'.  Stop.

This is because the source files are not correctly added into the
build.  Fix this, make the makefile match with pmie (as discussed),
remove a no-longer-needed header, and make some cosmetic changes in
the name of code consistency in rtime.c code.