improved error handling - pmGetContextHostName() usage
In a drive-by code review, I spotted a use of pmGetContextHostName()
with poor error handling ... an error here returns an empty string
which may be OK, but when it is not, we should die right there,
not hope vainly that some latter check will detect this and report
something that is not obviously related.

Checked all the code and found several dodgey uses in a similar
vein, so hardened 'em all up.