pmlogger_daily - fix problem with pmproxy connections
If the pmlogger control file contains a "host" of the form
"pmcd_host@some_pmproxy_host", then pmlogger_check works as
expected, but pmlogger_daily fails to restart the new pmlogger each
night, leaving it to pmlogger_check to fix the next time it runs.

The problem was really in pmnewlog ... this commit fixes it.