libpcp/src/check-statics: simplify, pessify
check-statics has been in some cases overly and in others underly
sensitive.

In the overly sensitive category, it's too sensitive to compilers'
choice of section placement and for static data and suffixing due to
function-static / optimization.  It seems safe to accept the symbols
with any of the usual prefixes and suffixes.

In the underly sensitive category, check-statics could exit with a
success rc in a variety of conditions, such as being interrupted, or a
failure of an inferior command, or even a script syntax error.  Let's
flip the $sts default around and add a 'set -e' to become very
conservative with respect to environmental catastrophes.

kenj corrected/tested on mac/*bsd/suse/debian
nathans corrected/tested on solaris
fche retested on plain old linux