Export pmgetopt_r interface for use by collectors and similar
This commit exposes the existing pmgetopt_r libpcp interface
so that it can be called directly.  This is useful for tools
that do not and will not ever want to use the standard option
processing code, and where dealing with that gets in the way.
This is the case for tools like dbpmda, collector tools, and
most things that are not PMAPI monitor tools, really.

Several more tools are brought into the long-options fold and
make use of this directly - dbpmda, pmconfig, pmfind, pmlock.

In addition, pmval and pminfo are updated to use a slightly
neater way of handling their options that conflict with those
in the standard set.