Split python PMAPI test into live and archive variants
Looking to extend the python PMAPI test, I noticed that
the archive component was already failing but that the
test still passes.  This is because both live & archive
modes send output into $seq.full, and both grep on "OK"
in that output file to verify pass.

So, if live mode passed, the test always passed.  Sadly,
archive mode throws its toys^W^Wan exception part way
through - the test should have been failing all along.

Fix the test issue by splitting into two tests.  This
makes it simpler, faster, workier, and debugging should
be easier now.  Test 707 is still failing though - Stan
will hopefully get to that shortly, else I will circle
back, sneak up on it when its not looking, and fix it
(the fix is to create an archive with all the metrics
that the test needs, commit it and use it in 707).

Also, added in a pmArchiveEnd check while I was there.