Fix test qa/707 by using a custom archive
Test 707 was failing since it was recently cleft in
two to separate out live and archive modes of test,
exposing a testing bug where the latter could never
fail.

This commit now fixes test 707, via addition of a new
archive (qa/src/pyapi) containing all of the metrics
it is expecting (missing sample.bin in particular was
the root cause of the earlier failure).