qa/new - adjustments for new ## convention in groups file
In the groups file, lines with prefixed with a single # are
comments that are interpreted as one-line descriptions that the
new script looks for if the user asks for a list of groups at
the end of the creation of a new qa test.

Lines with prefixed with ## are also comments, but are ignored
for the purpose of finding the one-line descriptions.  Block
(multi-line) descriptions of a group are the intended use of the
(newly introduced) ## prefix.