Rework the source tarball generation to use git tools
Previously, we generated the source tarball from macros in
the various makefiles.  Now, we generate the source tarball
beforehand based on the (active) contents of the repository
as git sees it.

This means we no longer need to worry about whether or not
new files are in the source tarball, and also means we can
rework the local-git-rpm-build (so not Fedora/EPEL/RHEL) to
not perform the entire build twice.

We must still run the configure phase twice (for rpm and deb
packaging) so spec and debian/control files are generated,
but that is where the doubling-up ends.