Initial Python script support for auto-command-line-parsing
Support for long options and automatic PCP option parsing on
the command lines of python monitor tools.  A new pmOptions
class is added to the pmapi python module, and a classmethod
on the pmContext class allows it to be injected into the new
context creation process.

Two test scripts are also added to exercise the python APIs,
exercised via new tests qa/739 and qa/741.  The former looks
at specifics of the pmOptions class, the latter ensures that
this new class interfaces correctly with the PMAPI pmContext
creation mechanism.