libpcp_pmda - slow start PMDA changes - part 1 (PMDAs written in C)
Allow PMDAs to call pmdaConnect() earlier.

This reduces the risk of a slow starting PMDA being killed with a
timeout in the credentials handshake that pmcd starts as soon as
a daemon PMDA is launched.

There is little code change here, just additional internal state
and checking to make sure the pmdaInterface structure has been
properly setup when pmdaConnect is called.

The rest is QA (qa/799 is new) and more concise wording for the
pmdaConnect(3) man page.

The C version of the simple PMDA has been changed to show how
pmdaConnect() can be called earlier in the PMDA's initialization.