RHBZ1044297: improve rpm-pmda concurrency complications
A review with rhbz1044297 in mind indicated a few possible problems
with the way rpm-pmda was using threads.  The main problem is that the
inotify_thread could run concurrently with the one-shot indom_thread,
which could lead to all kinds of mayhem, because the librpm api
functions in rpm_update_cache() are called without concurrency
protection.  We eliminate the indom-thread entirely, instead
triggering an initial round of rpm-cache population at the
head of inotify_thread.

Part of the trigger appeared to be rpmReadConfigFiles(), which is
highly stateful inside librpm.  We also now preclude multiple
calls to that function with a simple static flag.

Smaller problems included lack of error protection in the inotify
setup, and unnecessary processing of the received inotify events.
(If we get any events, it means we must refresh.)

The qa/750 test case's pmsleep fudge-factor was raised yet higher to
account for machines with a large rpm database.  There is an inherent
race here between system rpm-database updates and the rpmpmda
listener.  We might be able to solve more robustly polling the
rpm.refresh.count value, but this is left for the future.