qa/763 & slow PMDA - slow starting PMDA
The slow PMDA shows how connect_pmcd() may be used to better
handle slow startup for Perl PMDAs ... the use of connect_pmcd()
is optional to allow tha passing an failing cases to be explored.
The PMDA will also do slow fetches, but the library support for
this is not yet done.

qa/763 exercises the slow PMDA in the passing and failing cases.