qa/841 (new) & slow PMDA (Python version) - slow start PMDA
The Python version of the slow PMDA shows how self.connect_pmcd()
may be used to better handle slow startup for Python PMDAs ... the
use of self.connect_pmcd() is optional to allow the passing and
failing cases to be explored.  The PMDA will also do slow fetches,
but the library support for this is not yet done.

qa/841 exercises the Python version of the slow PMDA in the passing
and failing cases.  This is the dual of qa/763 that exercise the
Perl version of the slow PMDA.