src/include/builddefs.in - remove diagnostic fluff
Not sure when the appending to /tmp/eek.raw was added, but this is
totally useless ... and fails if /tmp/eek.raw exists and is owned
by someone else (root in my failing case).

Given the name of the file, I am sure I added the diagnostic without
review, so I'll remove it without review.