src/pmcd/pmdaproc.sh - allow Install/Remove from any directory
When the Perl and Python PMDA support was added, it introduced
a restriction that these PMDAs can only be installed from a
directory below $PCP_VAR_DIR/pmdas ... this is quite unnecessary.

Remove that restriction.

Also promote $pmda_dir (that was being used locally in pmdaSetup to
the global scope so that an Install/Remove script can call source
pmdaproc.sh and then change $pmda_dir to be some other absolute
pathname before calling pmdaSetup.  This would only be useful
if your PMDA files for some bizarre reason live in a different
directory to the one where the Install/Remove scripts are being
run ... but does allow reversion to the previous behaviour on
the off chance that some PMDA was depending upon it.