Update test 750 to more robustly detect an installed rpm
Test qa/750 and common.rpm are updated to remove the increasing
sleep-waiting-for-rpmdb-refresh interval, and replace it with a
check on the presence/absence (depending on whether we are doing
addition/removal) of a named package in the rpm instance domain.

As an added bonus for folks running QA often, the test now runs
quite a bit more quickly in the case where the rpm refresh takes
only a short time (which appears to be the normal case, locally
at least).