Resolve missed change-does-not-match-surrounding-code in pmdarpm
A recent threading fix to pmdarpm introduced an exciting new coding
style into this file, contrary to HACKING Conventions.  On review,
some were removed, but others overlooked.  This fixes the remaining
ones so that this code does not start to look like Frankenstein, in
the long term (too late, you may argue ;) ... but at least we try!)