Provide a python pmLookupName API mode allowing partial failure
The python wrapper for pmLookupName takes a hard-line approach to
metric name resolution - if any are missing, an exception will be
thrown, and any names which did resolve are tossed in the bin.

The underlying C API is designed to cater for missing metrics and
tools use this facility surprisingly often.  It allows for cross-
platform quirks to be elegantly handled, and other fallback modes
(eg. where two similar metrics exist, and a prefered one might be
unavailable).

This API extension is tackled in a backwards-compatible way, by
using an optional second argument to allow the "relaxed" handling
of missing metric names.

Tests qa/702 and qa/707 now exercise this little addition, via an
extension to the qa/src/test_pcp.python script which they share.