Add a Makepkgs mode allowing for absence of GNU tar
If we have no GNU tar, we can still stumble on with building the
packages via Makepkgs, thanks to git-archive.  Then downside of
this mode is that it does not pick up not-yet-committed changes.
So its added as a fallback mode, and is used only when a GNU tar
binary cannot be found.