libpcp pmns - fix obscure memory leak
To see this you'd need to be using PM_CONTEXT_LOCAL for a DSO
PMDA with dynamic metrics and then call pmGetChildren() or
pmGetChildrenStatus() and the metric at the root of the search
needs to be a non-leaf with more than zero children.

This corner case is exposed by qa/957.