Support for pcp(1) running system performance scripts
Add in support for pcp(1) to be used to invoke system tools
as scripts (like the existing pcp-free, pcp-uptime scripts)
with automatic argument passing/parsing.  So, no support is
needed in the scripts, further simplifying them.  Man page
updates and qa/900 is added to exercise this.

The -p/-P option switch in pcp(1) is to facilitate standard
option usage across all tools (the pmie stats are so rarely
used this is not expected to be problem in practice).

In addition, a pcp-numastat(1) script is added, along with
pmlogconf spec, man page, qa/743 and initial archives that
exercise it (thanks Ken).

With this third script addition, the python options support
is looking fairly solid.  The last little tweak made here is
to automatically enable the archive boundary options, in the
library, as it appears all scripts will want to use this by
default.