python pmapi.c: Don't crash on exceptions from options_callback()
In case the python routine being invoked exits with an exception
(result is NULL), we shouldn't try to Py_DECREF the bad boy.

For that matter, we shouldn't clear such exceptions with PyErr_Print()
at all, but pass them through to the calling python VM.  This cleanup
is left for later.