Add a pmgetopt(1) command for shell scripts long option parsing
Inspired by getopt(1), this commit adds a pmgetopt(1) for PCP shell
scripts to acquire cross-platform long option argument handling, in
addition to their existing short option forms.  Its output matches
closely with the quoted variant of the getopt(1) output, however it
differs significantly wrt configuration.  It also differs in that
it can be asked to call on pmUsageMessage(3) to generate the usage
message for the script.

Instead of an arcane combination of command line options to specify
both long and short options, a usage-message-alike configuration
syntax is implemented.  This aids conversion of our existing shell
scripts, since they have usage messages already.  It does imply the
use of temporary files for configuration but we make such extensive
use of such files already that I figure thats not a big problem for
PCP (though it likely would have been for the folks who originally
wrote the getopt(1) command).  This also gives opportunity to sneak
in short-cuts for specifying the "standard" PCPIntro(1) options.

A man page and test qa/978 are added to document and test it.  Also
the pcp(1) and pmlogmv(1) scripts have been initially converted to
demonstrate and further exercise its use.