Refactor include/pmtime.h working towards tree unification
One issue with reunifying pcp and pcp-gui is the presence
of pmtime.h below the include directory in both trees.  It
has quite different contents in both places, although from
a high level its all about time control clients.

Its installed only in the pcp tree, so we'll go with that
one as golden.  Move the gui variant into Qmc (Qt Metrics
Class) library, and do a better job of sharing some of the
code backing it in the process.