Add -d option into qa/src/interp1.c for new-style output
Commit 068ab84d2 has accidentally regressed test qa/091 by
changing the output that interp1.c produces, which it was
relying being a certain way.

Resolve this by adding a new option to get the new output,
and keep on using the old output in existing tests.  Quite
probably test 091 could be reworked to use the new output,
but going for the simplest fix at this stage.