Implement pmInDom_build and pmID_build using memcpy
While merging pcp-gui into pcp, the optimisation levels at
which the Qt test programs are built now follows the rest of
the build and is no longer forced to "debug" builds only.

This has uncovered a problem in the casting mechanism used
by pmID_build and pmInDom_build in impl.h, which ultimately
can result in a pmID of 0.0.0 (or pmInDom of 0.0) being set
due to the breaking of strict-aliasing rules.  QA test 1015,
which uses qmc_desc.cpp and passes the QmcDesc constructor a
PMID from pmID_build, picked up the problem.

I tossed up going with an ifdef __cplusplus variant of these
two helper interfaces, but in the end went with the simplest
approach of always using memcpy.