Fix error handling in pcp(1) when metric source unavailable
Resolve a couple of issues in the handling of pmNewContext failure
in pcp(1) script - in live mode, we were search pminfo stdout for
a failure message which was on stderr; in archive mode we had some
spurious additional warnings from pmdumplog reading the label.

Add test qa/830 exercising the correct behaviour for these cases.